Get multiserver basically working.
Add multiserver support to the orphan server adopter.

Refactor orphan server adopter. It has a delegate that knows about windows and sessions, but it does not.

Add support for "preemptive waiting". This is because SIGCHLD appears to be broken on macOS (https://stackoverflow.com/questions/59123366). When you kill a process by sending it a signal, you follow up with a preemptive wait. Such a process is marked as "will terminate" and will no longer be reported as a child. There are still some issues to work out here. If it does not terminate, then it will just live forever and after relaunching the app you won't see it again.

Handle losing the connection to the multiserver.

Handshake after attaching.

Add missing break statements to waitError:

Remove debugging code from server.

Flesh out iTermMultiServerConnection and move it to its own file.

Don't spawn multiple servers with same socket.

Send Wait requests.

Fix dangling pointer in iTermMultiServerProtocolParseMessageFromServer